-- imported EvoStream patches and fixes
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@745 01689507-d990-df11-9d7d-0017a411d416

+++ sources/common/src/platform/windows/strptime.cpp
@@ -104,14 +104,12 @@ static int
 match_string (const char **buf, const char **strs)
 {
     int i = 0;
-
     for (i = 0; strs[i] != NULL; ++i) {
-	int len = strlen (strs[i]);
-
-	if (strncasecmp (*buf, strs[i], len) == 0) {
-	    *buf += len;
-	    return i;
-	}
+		int len = (int)strlen (strs[i]);
+		if (strncasecmp (*buf, strs[i], len) == 0) {
+			*buf += len;
+			return i;
+		}
     }
     return -1;
 }
